+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/RxDocumentServiceRequest.java
@@ -221,9 +221,6 @@ public class RxDocumentServiceRequest implements Cloneable {
                     .fromHeader(this.headers.get(HttpConstants.HttpHeaders.PARTITION_KEY_RANGE_ID));
         }
 
-        if (StringUtils.isNotEmpty(this.headers.get(HttpConstants.HttpHeaders.SESSION_TOKEN))) {
-            this.originalSessionToken = this.headers.get(HttpConstants.HttpHeaders.SESSION_TOKEN);
-        }
     }
 
     private RxDocumentServiceRequest(DiagnosticsClientContext clientContext,